Migrate feature module plugin to HGP
