+++ app/feature/feature-chat/build.gradle.kts
@@ -4,10 +4,10 @@ hedvig {
   }
   serialization()
   compose()
+  featureModule()
 }
 plugins {
   id("hedvig.gradle.plugin")
-  id("hedvig.android.feature")
   id("hedvig.android.library")
 }
