+++ app/feature/feature-forever/build.gradle.kts
@@ -1,10 +1,10 @@
 hedvig {
   serialization()
   compose()
+  featureModule()
 }
 plugins {
   id("hedvig.gradle.plugin")
-  id("hedvig.android.feature")
   id("hedvig.android.library")
 }
