+++ app/shared/forever-ui/build.gradle.kts
@@ -5,7 +5,6 @@ hedvig {
 }
 plugins {
   id("hedvig.gradle.plugin")
-  id("hedvig.android.feature")
   id("hedvig.android.library")
 }
