+++ app/shared/tier-comparison/build.gradle.kts
@@ -5,7 +5,6 @@ hedvig {
 }
 plugins {
   id("hedvig.gradle.plugin")
-  id("hedvig.android.feature")
   id("hedvig.android.library")
   id("kotlin-parcelize")
 }
